+++ proto/v1/v1/setting_service.proto
@@ -315,7 +315,7 @@ message WorkspaceApprovalSetting {
     // When source is CREATE_DATABASE, support: resource.environment_id, resource.project_id, resource.db_engine, resource.database_name
     // When source is EXPORT_DATA, support: resource.environment_id, resource.project_id, resource.db_engine, resource.database_name, resource.schema_name, resource.table_name
     // When source is REQUEST_ROLE, support: resource.project_id, request.expiration_days, request.role
-    // When source is REQUEST_ACCESS, support: resource.environment_id, resource.project_id, request.unmask, request.data_export
+    // When source is REQUEST_ACCESS, support: resource.environment_id, resource.project_id, resource.db_engine, resource.database_name, resource.schema_name, resource.table_name, request.unmask, request.data_export
     //
     // For examples:
     // resource.environment_id == "prod" && statement.affected_rows >= 100
